reporter: Split ReporterTest into separate tests per output format
To align with existing format-specific tests and to allow tests to run
in parallel.
